Refactor caching out of ExecutorContext plus some tidying up
