+++ src/wasm-lib/kcl-test-server/src/lib.rs
@@ -151,7 +151,7 @@ async fn handle_request(req: hyper::Request<Body>, state3: Arc<ServerState>) -> 
 /// KCL errors (from engine or the executor) respond with HTTP Bad Gateway.
 /// Malformed requests are HTTP Bad Request.
 /// Successful requests contain a PNG as the body.
-async fn snapshot_endpoint(body: Bytes, state: ExecutorContext) -> Response<Body> {
+async fn snapshot_endpoint(body: Bytes, ctxt: ExecutorContext) -> Response<Body> {
     let body = match serde_json::from_slice::<RequestBody>(body.as_ref()) {
         Ok(bd) => bd,
         Err(e) => return bad_request(format!("Invalid request JSON: {e}")),
@@ -164,11 +164,11 @@ async fn snapshot_endpoint(body: Bytes, state: ExecutorContext) -> Response<Body
     };
 
     eprintln!("Executing {test_name}");
-    let mut exec_state = ExecState::new(&state.settings);
+    let mut exec_state = ExecState::new(&ctxt.settings);
     // This is a shitty source range, I don't know what else to use for it though.
     // There's no actual KCL associated with this reset_scene call.
-    if let Err(e) = state
-        .reset_scene(&mut exec_state, kcl_lib::SourceRange::default())
+    if let Err(e) = ctxt
+        .send_clear_scene(&mut exec_state, kcl_lib::SourceRange::default())
         .await
     {
         return kcl_err(e);
@@ -176,7 +176,7 @@ async fn snapshot_endpoint(body: Bytes, state: ExecutorContext) -> Response<Body
     // Let users know if the test is taking a long time.
     let (done_tx, done_rx) = oneshot::channel::<()>();
     let timer = time_until(done_rx);
-    let snapshot = match state.execute_and_prepare_snapshot(&program, &mut exec_state).await {
+    let snapshot = match ctxt.execute_and_prepare_snapshot(program.into(), &mut exec_state).await {
         Ok(sn) => sn,
         Err(e) => return kcl_err(e),
     };
